Add writes, use an env for mutation tokens.
This commit combines two changes which probably should be separate, but
it's early in the example to worry about atomicity of changes.

Motivation:
-----------
There are some situations where write availability is of primary concern.
In that case, we want to ensure that not only do we switch to the
alternate cluster, but also handle any writes that cannot be completed
against the primary cluster.

Modifications:
--------------
This change adds a new example that shows how writes can be achived on an
alternate cluster from an appliction.  It does so by extending the XCBucket
interface with a way to upsert() to the alternate cluster.

Results:
--------
No writes are lost.  FallbackBucket will continue to perform failover
based on the event bus signaling, but this is just the most complex example.
Failures may also be handled heruistically or by some other external
declaration method.